--- core_backend/app/voice_api/__init__.py
@@ -1,3 +1,0 @@
-from .routers import router
-
-__all__ = ["router"]
